move fixture out
